refactor(transport): share declaration handle lifecycle (#11)
- Centralize subscription and queryable ownership transitions\n- Keep backend-specific native handle cleanup isolated\n- Eliminate duplicated zenoh and stub lifecycle implementations
