+++ CMakeLists.txt
@@ -61,6 +61,7 @@ add_library(sitos STATIC
   src/in_memory_engine.cpp
   src/logging.cpp
   src/storage_node.cpp
+  src/transport/declaration_handle_lifecycle.cpp
 )
 add_library(sitos::sitos ALIAS sitos)
 target_include_directories(sitos PUBLIC
